// types/kyc.ts
export type NextOfKinType = {
    fullName: string
    phone1: string
    phone2: string
    email1: string
    email2: string
    address: string
    gender: string
    relationship: string
  }
  
  export type UpdateNextOfKinType = {
      fullName?: string
      phone1?: string
      phone2?: string
      email1?: string
      email2?: string
      address?: string
      gender?: string
      relationship?: string
    }
    